feat(users): add CRUD capability and login controller
To create an admin user req.body must provide an apikey
and that apikey must be generated for the system before.

The updateUser method may recieve an object with any qty
of keys and the controller will attempt to update those
who exist in the userModel schema.

for example:

if the controller recieve { "name": "xxxx", email: "xxx" }
then will update the name and the email of the user.